# api/openapi.yaml
openapi: 3.0.3
info:
  title: happyDeliver API
  description: Email Deliverability Testing Platform API
  version: 0.1.0
  contact:
    name: happyDomain team
    url: https://github.com/happyDomain/happydeliver
    email: contact+api@happydomain.org
  license:
    name: GNU Affero General Public License v3.0 or later
    url: https://spdx.org/licenses/AGPL-3.0-or-later.html
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.example.com/api
    description: Production server
...
paths:
  /test:
    post:
      tags:
        - tests
      summary: Create a new deliverability test
      description: Generates a unique test email address for sending test emails. No database record is created until an email is received.
      operationId: createTest
      responses:
        '201':
          description: Test email address generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /test/{id}:
    get:
      tags:
        - tests
      summary: Get test status
      description: Check if a report exists for the given test ID (base32-encoded). Returns pending if no report exists, analyzed if a report is available.
      operationId: getTest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
            description: Base32-encoded test ID (with hyphens)
      responses:
        '200':
          description: Test status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tests:
    get:
      tags:
        - tests
      summary: List all tests
      description: Returns a paginated list of test summaries with scores and grades. Can be disabled via server configuration.
      operationId: listTests
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of items to skip
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum number of items to return
      responses:
        '200':
          description: List of test summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestListResponse'
        '403':
          description: Test listing is disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /report/{id}:
    get:
      tags:
        - reports
      summary: Get detailed report
      description: Retrieve comprehensive deliverability analysis report
      operationId: getReport
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
            description: Base32-encoded test ID (with hyphens)
      responses:
        '200':
          description: Report retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /report/{id}/raw:
    get:
      tags:
        - reports
      summary: Get raw annotated email
      description: Retrieve the original email with headers added by filters
      operationId: getRawEmail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
            description: Base32-encoded test ID (with hyphens)
      responses:
        '200':
          description: Raw email retrieved successfully
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Email not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /report/{id}/reanalyze:
    post:
      tags:
        - reports
      summary: Reanalyze email and regenerate report
      description: Re-run the analysis on the stored raw email to regenerate the report with the latest analyzer version. This is useful after analyzer improvements or bug fixes.
      operationId: reanalyzeReport
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
            description: Base32-encoded test ID (with hyphens)
      responses:
        '200':
          description: Report regenerated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '404':
          description: Email not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error during reanalysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /domain:
    post:
      tags:
        - tests
      summary: Test a domain's email configuration
      description: Analyzes DNS records (MX, SPF, DMARC, BIMI) for a domain without requiring an actual email to be sent. Returns results immediately.
      operationId: testDomain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainTestRequest'
      responses:
        '200':
          description: Domain test completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainTestResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /blacklist:
    post:
      tags:
        - tests
      summary: Check an IP address against DNS blacklists
      description: Tests a single IP address (IPv4 or IPv6) against configured DNS-based blacklists (RBLs) without requiring an actual email to be sent. Returns results immediately.
      operationId: checkBlacklist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlacklistCheckRequest'
      responses:
        '200':
          description: Blacklist check completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlacklistCheckResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Test:
      type: object
      required:
        - id
        - email
        - status
      properties:
        id:
          type: string
          pattern: '^[a-z0-9-]+$'
          description: Unique test identifier (base32-encoded with hyphens)
          example: "krfwg4z-amrqw4z-zmorsw2-djmfzgk-3a"
        email:
          type: string
          format: email
          description: Unique test email address
          example: "test-krfwg4z-amrqw4z-zmorsw2-djmfzgk-3a@example.com"
        status:
          type: string
          enum: [pending, analyzed]
          description: Current test status (pending = no report yet, analyzed = report available)
          example: "analyzed"
    TestResponse:
      type: object
      required:
        - id
        - email
        - status
      properties:
        id:
          type: string
          pattern: '^[a-z0-9-]+$'
          description: Unique test identifier (base32-encoded with hyphens)
          example: "krfwg4z-amrqw4z-zmorsw2-djmfzgk-3a"
        email:
          type: string
          format: email
          example: "test-krfwg4z-amrqw4z-zmorsw2-djmfzgk-3a@example.com"
        status:
          type: string
          enum: [pending]
          example: "pending"
        message:
          type: string
          example: "Send your test email to the address above"
    Report:
      type: object
      required:
        - id
        - test_id
        - score
        - grade
        - created_at
      properties:
        id:
          type: string
          pattern: '^[a-z0-9-]+$'
          description: Report identifier (base32-encoded with hyphens)
        test_id:
          type: string
          pattern: '^[a-z0-9-]+$'
          description: Associated test ID (base32-encoded with hyphens)
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: Overall deliverability score as percentage (0-100)
          example: 85
        grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade representation of the score (A+ is best, F is worst)
          example: "A"
        summary:
          $ref: '#/components/schemas/ScoreSummary'
        authentication:
          $ref: '#/components/schemas/AuthenticationResults'
        spamassassin:
          $ref: '#/components/schemas/SpamAssassinResult'
        rspamd:
          $ref: '#/components/schemas/RspamdResult'
        dns_results:
          $ref: '#/components/schemas/DNSResults'
        blacklists:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/BlacklistCheck'
          description: Map of IP addresses to their blacklist check results (array of checks per IP)
          example:
            "192.0.2.1":
              - rbl: "zen.spamhaus.org"
                listed: false
              - rbl: "bl.spamcop.net"
                listed: false
        whitelists:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/BlacklistCheck'
          description: Map of IP addresses to their DNS whitelist check results (informational only)
          example:
            "192.0.2.1":
              - rbl: "list.dnswl.org"
                listed: false
              - rbl: "swl.spamhaus.org"
                listed: false
        content_analysis:
          $ref: '#/components/schemas/ContentAnalysis'
        header_analysis:
          $ref: '#/components/schemas/HeaderAnalysis'
        raw_headers:
          type: string
          description: Raw email headers
        created_at:
          type: string
          format: date-time
    ScoreSummary:
      type: object
      required:
        - dns_score
        - dns_grade
        - authentication_score
        - authentication_grade
        - spam_score
        - spam_grade
        - blacklist_score
        - blacklist_grade
        - header_score
        - header_grade
        - content_score
        - content_grade
      properties:
        dns_score:
          type: integer
          minimum: 0
          maximum: 100
          description: DNS records score (in percentage)
          example: 42
        dns_grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade representation of the score (A+ is best, F is worst)
          example: "A"
        authentication_score:
          type: integer
          minimum: 0
          maximum: 100
          description: SPF/DKIM/DMARC score (in percentage)
          example: 28
        authentication_grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade representation of the score (A+ is best, F is worst)
          example: "A"
        spam_score:
          type: integer
          minimum: 0
          maximum: 100
          description: Spam filter score (SpamAssassin + rspamd combined, in percentage)
          example: 15
        spam_grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade representation of the score (A+ is best, F is worst)
          example: "A"
        blacklist_score:
          type: integer
          minimum: 0
          maximum: 100
          description: Blacklist check score (in percentage)
          example: 20
        blacklist_grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade representation of the score (A+ is best, F is worst)
          example: "A"
        header_score:
          type: integer
          minimum: 0
          maximum: 100
          description: Header quality score (in percentage)
          example: 9
        header_grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade representation of the score (A+ is best, F is worst)
          example: "A"
        content_score:
          type: integer
          minimum: 0
          maximum: 100
          description: Content quality score (in percentage)
          example: 18
        content_grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade representation of the score (A+ is best, F is worst)
          example: "A"
...
    AuthenticationResults:
      type: object
      properties:
        spf:
          $ref: '#/components/schemas/AuthResult'
        dkim:
          type: array
          items:
            $ref: '#/components/schemas/AuthResult'
        dmarc:
          $ref: '#/components/schemas/AuthResult'
        bimi:
          $ref: '#/components/schemas/AuthResult'
        arc:
          $ref: '#/components/schemas/ARCResult'
        iprev:
          $ref: '#/components/schemas/IPRevResult'
        x_google_dkim:
          $ref: '#/components/schemas/AuthResult'
          description: Google-specific DKIM authentication result (x-google-dkim)
        x_aligned_from:
          $ref: '#/components/schemas/AuthResult'
          description: X-Aligned-From authentication result (checks address alignment)
...
    SpamAssassinResult:
      type: object
      required:
        - score
        - required_score
        - is_spam
        - test_details
      properties:
        deliverability_score:
          type: integer
          minimum: 0
          maximum: 100
          description: SpamAssassin deliverability score (0-100, higher is better)
          example: 80
        deliverability_grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade for SpamAssassin deliverability score
          example: "B"
        version:
          type: string
          description: SpamAssassin version
          example: "SpamAssassin 4.0.1"
        score:
          type: number
          format: float
          description: SpamAssassin spam score
          example: 2.3
        required_score:
          type: number
          format: float
          description: Threshold for spam classification
          example: 5.0
        is_spam:
          type: boolean
          description: Whether message is classified as spam
          example: false
        tests:
          type: array
          items:
            type: string
          description: List of triggered SpamAssassin tests
          example: ["BAYES_00", "DKIM_SIGNED"]
        test_details:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SpamTestDetail'
          description: Map of test names to their detailed results
          example:
            BAYES_00:
              name: "BAYES_00"
              score: -1.9
              description: "Bayes spam probability is 0 to 1%"
            DKIM_SIGNED:
              name: "DKIM_SIGNED"
              score: 0.1
              description: "Message has a DKIM or DK signature, not necessarily valid"
        report:
          type: string
          description: Full SpamAssassin report
...
    RspamdResult:
      type: object
      required:
        - score
        - threshold
        - is_spam
        - symbols
      properties:
        deliverability_score:
          type: integer
          minimum: 0
          maximum: 100
          description: rspamd deliverability score (0-100, higher is better)
          example: 85
        deliverability_grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade for rspamd deliverability score
          example: "A"
        score:
          type: number
          format: float
          description: rspamd spam score
          example: -3.91
        threshold:
          type: number
          format: float
          description: Score threshold for spam classification
          example: 15.0
        action:
          type: string
          description: rspamd action (no action, add header, rewrite subject, soft reject, reject)
          example: "no action"
        is_spam:
          type: boolean
          description: Whether message is classified as spam (action is reject or soft reject)
          example: false
        server:
          type: string
          description: rspamd server that processed the message
          example: "rspamd.example.com"
        symbols:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SpamTestDetail'
          description: Map of triggered rspamd symbols to their details
          example:
            BAYES_HAM:
              name: "BAYES_HAM"
              score: -1.9
              params: "0.02"
        report:
          type: string
          description: Full rspamd report (raw X-Spamd-Result header)
...
    BlacklistCheck:
      type: object
      required:
        - rbl
        - listed
      properties:
        rbl:
          type: string
          description: RBL/DNSBL name
          example: "zen.spamhaus.org"
        listed:
          type: boolean
          description: Whether IP is listed
          example: false
        response:
          type: string
          description: RBL response code or message
          example: "127.0.0.2"
        error:
          type: string
          description: RBL error if any
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "not_found"
        message:
          type: string
          description: Human-readable error message
          example: "Test not found"
        details:
          type: string
          description: Additional error details
    DomainTestRequest:
      type: object
      required:
        - domain
      properties:
        domain:
          type: string
          pattern: '^[a-zA-Z0-9]([a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(\.[a-zA-Z0-9]([a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$'
          description: Domain name to test (e.g., example.com)
          example: "example.com"
    DomainTestResponse:
      type: object
      required:
        - domain
        - score
        - grade
        - dns_results
      properties:
        domain:
          type: string
          description: The tested domain name
          example: "example.com"
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: Overall domain configuration score (0-100)
          example: 85
        grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade representation of the score
          example: "A"
        dns_results:
          $ref: '#/components/schemas/DNSResults'
    BlacklistCheckRequest:
      type: object
      required:
        - ip
      properties:
        ip:
          type: string
          description: IPv4 or IPv6 address to check against blacklists
          example: "192.0.2.1"
          pattern: '^([0-9]{1,3}\.){3}[0-9]{1,3}$|^([0-9a-fA-F]{0,4}:){7}[0-9a-fA-F]{0,4}$|^::([0-9a-fA-F]{0,4}:){0,6}[0-9a-fA-F]{0,4}$|^([0-9a-fA-F]{0,4}:){1,6}:([0-9a-fA-F]{0,4}:){0,5}[0-9a-fA-F]{0,4}$'
    BlacklistCheckResponse:
      type: object
      required:
        - ip
        - blacklists
        - listed_count
        - score
        - grade
      properties:
        ip:
          type: string
          description: The IP address that was checked
          example: "192.0.2.1"
        blacklists:
          type: array
          items:
            $ref: '#/components/schemas/BlacklistCheck'
          description: List of blacklist check results
        listed_count:
          type: integer
          description: Number of blacklists that have this IP listed
          example: 0
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: Blacklist score (0-100, higher is better)
          example: 100
        grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade representation of the score
          example: "A+"
        whitelists:
          type: array
          items:
            $ref: '#/components/schemas/BlacklistCheck'
          description: List of DNS whitelist check results (informational only)
    TestSummary:
      type: object
      required:
        - test_id
        - score
        - grade
        - created_at
      properties:
        test_id:
          type: string
          pattern: '^[a-z0-9-]+$'
          description: Test identifier (base32-encoded with hyphens)
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: Overall deliverability score (0-100)
        grade:
          type: string
          enum: [A+, A, B, C, D, E, F]
          description: Letter grade
        from_domain:
          type: string
          description: Sender domain extracted from the report
        created_at:
          type: string
          format: date-time
    TestListResponse:
      type: object
      required:
        - tests
        - total
        - offset
        - limit
      properties:
        tests:
          type: array
          items:
            $ref: '#/components/schemas/TestSummary'
        total:
          type: integer
          description: Total number of tests
        offset:
          type: integer
          description: Current offset
        limit:
          type: integer
          description: Current limit